refactor: Improve directory structure for ProductList and ProductItem components
